requirements.txt
flask
requests
beautifulsoup4
pydantic
openai
python-dotenv
pandas
pymongo